requirements/base.txt
django>=5.0,<6.0
psycopg2-binary>=2.9
python-decouple>=3.8
Pillow>=10.0
django-crispy-forms>=2.1
crispy-tailwind>=1.0
django-otp>=1.2
qrcode[pil]>=7.4
django-htmx>=1.17
python-magic>=0.4.27
whitenoise>=6.6
gunicorn>=21.2
celery>=5.3
redis>=5.0
django-celery-beat>=2.5
reportlab>=4.0
django-ratelimit>=4.1
